Updated pal_lld (GPIO low-level driver) fro BCM2835 to use atomic GPIOx_SET and GPIOx_CLR registers. This prevents issues when two threads simultaneously write to different pins on the same GPIO peripheral. What could happen was:
- Thread 1 reads latch (=GPIO state stored in memory)
- Thread 1 ORs that value to set bit x
- Context switch occurs
- Thread 2 reads latch
- Thread 2 ORs that value to set bit y
- Thread 2 writes the value to GPIO
- Context switch occurs
- Thread 2 writes its OR'ed value to GPIO, clearing bit y that was set
    by thread 2

By correctly using the GPIOx_SET and GPIOx_CLR hardware, this issue is avoided.


The pal_lld_readlatch implementation was removed since it cannot be guaranteed to work correctly if a GPIO is used from different threads. So, to avoid hard-to-find errors, the implementation was removed. Instead of the common read-modify-write approach, one should use palSetPort, palClearPort, or palWriteGroup, which correctly handle setting and clearing GPIO bits by means of the GPIOx_SET and GPIOx_CLR